SONAR-18638 Drop deprecated fields in Rules WS API.
Drop "effortToFixDescription" and replace with "gapDescription".
Drop "debtRemFnCoeff" and replace with "remFnGapMultiplier".
Drop "defaultDebtRemFnCoeff" and replace with "defaultRemFnGapMultiplier".
Drop "debtRemFnOffset" and replace with "remFnBaseEffort".
Drop "defaultDebtRemFnOffset" and replace with "defaultRemFnBaseEffort".
Correct the version when the field 'scope' was added.
Drop "debtOverloaded" and replace with "remFnOverloaded".
Deprecate "defaultDebtRemFnType" and "debtRemFnType" and replace with "defaultRemFnType" and "remFnType" accordingly.
Deprecate "defaultDebtRemFn" and "debtRemFn" and replace with "defaultRemFn" and "remFn" accordingly.